Change desrciptions array to description string
Since we're not working on different types of descriptions right now,
I'm just changing it to a single string called description for now.

Eventually, I think description should become an object of type Info. It
would contain short, long, and event-dependent descriptions or whatever
else we need. By making it an object, we could change how it works
without continually changing the definitions of all the classes.